Fix types in edltestutils.h
For the `struct unused`, the template parameter is receiving the type of
`sizeof`, which is not an `int`, and GCC 7 warns about data loss. The
correct type to use here is `std::size_t` as it is the defined return
type of `sizeof`. The macro `OE_UNUSED` compares `i` to `N`, so I have
made the types match.

In `array_compare`, we again took `int` even though it is a template
parameter used in an array access operator, which by definition should
be `std::size_t`. This is also true when we later iterate through the
array.